a program that prints the first 50 Fibonacci numbers, starting with 1 and 2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,20 +8,20 @@
  */
 void fibonacci(void)
 {
-	long long i = 1, j = 2, result = 0, n = 3;
+	int i = 1, j = 2, result = 0, n = 3;
 
-		printf("%lld, ", i);
-		printf("%lld, ", j);
+		printf("%d, ", i);
+		printf("%d, ", j);
 	while (n <= 50)
 	{
 		result = (i + j);
 		if (n == 50)
 		{
-			printf("%lld", result);
+			printf("%d", result);
 		}
 		else
 		{
-			printf("%lld, ", result);
+			printf("%d, ", result);
 		}
 		i = j;
 		j = result;
